+++ packages/~/moderations/lib/src/usecase/mark_moderation_as.ts
@@ -10,13 +10,15 @@ import { append_comment, type Comment_Type } from "../comment_message";
 
 export async function mark_moderation_as(
   {
-    pg,
     moderation,
+    pg,
+    reason,
     userinfo,
   }: {
     moderation: Pick<typeof schema.moderations.$inferSelect, "comment" | "id">;
-    userinfo: AgentConnect_UserInfo;
     pg: MonComptePro_PgDatabase;
+    reason: string;
+    userinfo: AgentConnect_UserInfo;
   },
   type: Comment_Type["type"],
 ) {
@@ -28,6 +30,7 @@ export async function mark_moderation_as(
   await update_moderation_by_id(moderation_id, {
     comment: append_comment(comment, {
       created_by: userinfo.email,
+      reason,
       type,
     }),
     moderated_by,