+++ packages/~/organizations/api/src/leaders/index.tsx
@@ -2,9 +2,11 @@
 
 import { zValidator } from "@hono/zod-validator";
 import env from "@~/app.core/config";
+import { button } from "@~/app.ui/button";
 import consola from "consola";
 import { Hono } from "hono";
 import lodash_sortby from "lodash.sortby";
+import { match, P } from "ts-pattern";
 import { z } from "zod";
 
 //
@@ -27,12 +29,22 @@ export default new Hono().get(
 
 async function List_Leaders({ siret }: { siret: string }) {
   const doc = await load_leaders({ siret });
-  if (!doc) return <>Pas de liste des dirigeants</>;
-  return (
-    <a class="fr-link" href={doc.url} rel="noopener noreferrer" target="_blank">
-      Liste des dirigeants
-    </a>
-  );
+  return match(doc)
+    .with({ url: P.string }, ({ url }) => (
+      <a
+        class={button({ className: "bg-white", size: "sm", type: "tertiary" })}
+        href={url}
+        rel="noopener noreferrer"
+        target="_blank"
+      >
+        Liste dirigeants associations
+      </a>
+    ))
+    .otherwise(() => (
+      <a class={button({ size: "sm", type: "tertiary" })}>
+        Pas de liste des dirigeants
+      </a>
+    ));
 }
 
 async function load_leaders({ siret }: { siret: string }) {