+++ packages/~/organizations/ui/src/info/Investigation.tsx
@@ -1,6 +1,7 @@
 //
 
 import { button } from "@~/app.ui/button";
+import { hx_urls } from "@~/app.urls";
 
 import type { Organization } from "@~/organizations.lib/entities/Organization";
 import type { JSX } from "hono/jsx";
@@ -13,45 +14,50 @@ type Props = JSX.IntrinsicElements["section"] & {
 
 //
 
-export function Investigation(props: Props) {
-  const { organization, ...section_props } = props;
+export async function Investigation(props: Props) {
+  const { organization } = props;
+  const hx_organizations_leaders_props =
+    await hx_urls.organizations.leaders.$get({
+      query: { siret: organization.siret },
+    });
 
   return (
-    <section {...section_props}>
-      <h4>🕵️ Enquête sur cette organisation</h4>
-
-      <ul class="list-none pl-0">
-        <li>
-          <a
-            href={`https://lannuaire.service-public.fr/recherche?where=${organization.cached_code_postal}&whoWhat=Mairie`}
-            class={button({ size: "sm", type: "tertiary" })}
-            rel="noopener noreferrer"
-            target="_blank"
-          >
-            Mairie sur Annuaire Service Public
-          </a>
-        </li>
-        <li>
-          <a
-            href={`https://lannuaire.service-public.fr/recherche?where=${organization.cached_code_postal}`}
-            class={button({ size: "sm", type: "tertiary" })}
-            rel="noopener noreferrer"
-            target="_blank"
-          >
-            Service sur Annuaire Service Public
-          </a>
-        </li>
-        <li>
-          <a
-            href={`https://data.education.gouv.fr/api/v2/catalog/datasets/fr-en-annuaire-education/records?where=siren_siret%3D${organization.siret}`}
-            class={button({ size: "sm", type: "tertiary" })}
-            rel="noopener noreferrer"
-            target="_blank"
-          >
-            Établissement sur l'annuaire Éducation Nationale
-          </a>
-        </li>
-      </ul>
-    </section>
+    <ul class="mt-5 w-full list-none bg-[#F6F6F6] p-3 [&_li]:inline-block">
+      <li>
+        <a
+          href={`https://lannuaire.service-public.fr/recherche?where=${organization.cached_code_postal}&whoWhat=Mairie`}
+          class={`${button({ size: "sm", type: "tertiary" })} mr-2 bg-white `}
+          rel="noopener noreferrer"
+          target="_blank"
+        >
+          Chercher la mairie associée
+        </a>
+      </li>
+      <li>
+        <a
+          href={`https://lannuaire.service-public.fr/recherche?where=${organization.cached_code_postal}`}
+          class={`${button({ size: "sm", type: "tertiary" })} mr-2 bg-white `}
+          rel="noopener noreferrer"
+          target="_blank"
+        >
+          Chercher les services publics associés
+        </a>
+      </li>
+      <li>
+        <a
+          href={`https://annuaire-entreprises.data.gouv.fr/dirigeants/${organization.siret.substring(0, 9)}`}
+          class={`${button({ size: "sm", type: "tertiary" })} mr-2 bg-white `}
+          rel="noopener noreferrer"
+          target="_blank"
+        >
+          Liste dirigeants - Annuaire entreprise API
+        </a>
+      </li>
+      <li>
+        <a class="" {...hx_organizations_leaders_props} hx-trigger="load">
+          <i class="text-center">Recherche des dirigeants...</i>
+        </a>
+      </li>
+    </ul>
   );
 }