+++ packages/~/users/ui/src/About/About.tsx
@@ -2,7 +2,7 @@
 
 import { z_email_domain } from "@~/app.core/schema/z_email_domain";
 import { CopyButton } from "@~/app.ui/button/components/copy";
-import { LocalTime } from "@~/app.ui/time/LocalTime";
+import { description_list } from "@~/app.ui/list";
 import { urls } from "@~/app.urls";
 import type { GetUserInfoOutput } from "@~/users.lib/usecase/GetUserInfo";
 
@@ -21,6 +21,8 @@ export function About({ user }: AboutProps) {
     <section>
       <h3>
         <a
+          class="bg-none"
+          target="_blank"
           href={
             urls.users[":id"].$url({ param: { id: user.id.toString() } })
               .pathname
@@ -29,55 +31,39 @@ export function About({ user }: AboutProps) {
           👨‍💻 Profile
         </a>
       </h3>
+      <dl class={description_list()}>
+        <dt>Email </dt>
+        <dd>
+          {user.email}{" "}
+          <CopyButton
+            class="ml-2"
+            text={user.email}
+            variant={{ size: "sm", type: "tertiary" }}
+          ></CopyButton>
+        </dd>
 
-      <ul class="list-none pl-0">
-        <li>
-          Email : <b>{user.email}</b>
-        </li>
-        <li>
-          Prénom : <b>{user.given_name}</b>
-        </li>
-        <li>
-          Nom : <b>{user.family_name}</b>
-        </li>
-        <li>
-          Téléphone : <b>{user.phone_number}</b>
-        </li>
-        <li>
-          Profession : <b>{user.job}</b>
-        </li>
-      </ul>
+        <dt>Domaine email </dt>
+        <dd>
+          {domain}{" "}
+          <CopyButton
+            class="ml-2"
+            text={domain}
+            variant={{ size: "sm", type: "tertiary" }}
+          ></CopyButton>
+        </dd>
 
-      <CopyButton text={user.email} variant={{ size: "sm", type: "tertiary" }}>
-        Copier l'email
-      </CopyButton>
-      <CopyButton text={domain} variant={{ size: "sm", type: "tertiary" }}>
-        Copier le domaine
-      </CopyButton>
+        <dt>Prénom </dt>
+        <dd>{user.given_name}</dd>
 
-      <details class="my-6">
-        <summary>Détails du profile</summary>
-        <ul>
-          <li>
-            id : <b>{user.id}</b>
-          </li>
-          <li>
-            Dernière connexion :{" "}
-            <b>
-              <LocalTime date={user.last_sign_in_at} />
-            </b>
-          </li>
-          <li>
-            Creation de compte :{" "}
-            <b>
-              <LocalTime date={user.created_at} />
-            </b>
-          </li>
-          <li>
-            Nombre de connection : <b>{user.sign_in_count}</b>
-          </li>
-        </ul>
-      </details>
+        <dt>Nom </dt>
+        <dd>{user.family_name}</dd>
+
+        <dt>Numéro de téléphone </dt>
+        <dd>{user.phone_number}</dd>
+
+        <dt>Profession </dt>
+        <dd>{user.job}</dd>
+      </dl>
     </section>
   );
 }